+++ src/main/webapp/resources/parsleyjs/src/parsley/ui.js
@@ -1,96 +1,153 @@
 import $ from 'jquery';
 import ParsleyUtils from './utils';
 
-var ParsleyUI = function (options) {
-  this.__class__ = 'ParsleyUI';
+var ParsleyUI = {};
+
+var diffResults = function (newResult, oldResult, deep) {
+  var added = [];
+  var kept = [];
+
+  for (var i = 0; i < newResult.length; i++) {
+    var found = false;
+
+    for (var j = 0; j < oldResult.length; j++)
+      if (newResult[i].assert.name === oldResult[j].assert.name) {
+        found = true;
+        break;
+      }
+
+    if (found)
+      kept.push(newResult[i]);
+    else
+      added.push(newResult[i]);
+  }
+
+  return {
+    kept: kept,
+    added: added,
+    removed: !deep ? diffResults(oldResult, newResult, true).added : []
+  };
 };
 
-ParsleyUI.prototype = {
-  listen: function () {
-    window.Parsley
-    .on('form:init',       (form ) => { this.setupForm (form ); } )
-    .on('field:init',      (field) => { this.setupField(field); } )
-    .on('field:validated', (field) => { this.reflow    (field); } )
-    .on('form:validated',  (form ) => { this.focus     (form ); } )
-    .on('field:reset',     (field) => { this.reset     (field); } )
-    .on('form:destroy',    (form ) => { this.destroy   (form ); } )
-    .on('field:destroy',   (field) => { this.destroy   (field); } );
+ParsleyUI.Form = {
 
-    return this;
+  _actualizeTriggers: function () {
+    this.$element.on('submit.Parsley', evt => { this.onSubmitValidate(evt); });
+    this.$element.on('click.Parsley', 'input[type="submit"], button[type="submit"]', evt => { this.onSubmitButton(evt); });
+
+    // UI could be disabled
+    if (false === this.options.uiEnabled)
+      return;
+
+    this.$element.attr('novalidate', '');
   },
 
-  reflow: function (fieldInstance) {
-    // If this field has not an active UI (case for multiples) don't bother doing something
-    if ('undefined' === typeof fieldInstance._ui || false === fieldInstance._ui.active)
+  focus: function () {
+    this._focusedField = null;
+
+    if (true === this.validationResult || 'none' === this.options.focus)
+      return null;
+
+    for (var i = 0; i < this.fields.length; i++) {
+      var field = this.fields[i];
+      if (true !== field.validationResult && field.validationResult.length > 0 && 'undefined' === typeof field.options.noFocus) {
+        this._focusedField = field.$element;
+        if ('first' === this.options.focus)
+          break;
+      }
+    }
+
+    if (null === this._focusedField)
+      return null;
+
+    return this._focusedField.focus();
+  },
+
+  _destroyUI: function () {
+    // Reset all event listeners
+    this.$element.off('.Parsley');
+  }
+
+};
+
+ParsleyUI.Field = {
+
+  _reflowUI: function () {
+    this._buildUI();
+
+    // If this field doesn't have an active UI don't bother doing something
+    if (!this._ui)
       return;
 
     // Diff between two validation results
-    var diff = this._diff(fieldInstance.validationResult, fieldInstance._ui.lastValidationResult);
+    var diff = diffResults(this.validationResult, this._ui.lastValidationResult);
 
     // Then store current validation result for next reflow
-    fieldInstance._ui.lastValidationResult = fieldInstance.validationResult;
+    this._ui.lastValidationResult = this.validationResult;
 
     // Handle valid / invalid / none field class
-    this.manageStatusClass(fieldInstance);
+    this._manageStatusClass();
 
     // Add, remove, updated errors messages
-    this.manageErrorsMessages(fieldInstance, diff);
+    this._manageErrorsMessages(diff);
 
     // Triggers impl
-    this.actualizeTriggers(fieldInstance);
+    this._actualizeTriggers();
 
     // If field is not valid for the first time, bind keyup trigger to ease UX and quickly inform user
-    if ((diff.kept.length || diff.added.length) && true !== fieldInstance._ui.failedOnce)
-      this.manageFailingFieldTrigger(fieldInstance);
+    if ((diff.kept.length || diff.added.length) && !this._failedOnce) {
+      this._failedOnce = true;
+      this._actualizeTriggers();
+    }
   },
 
   // Returns an array of field's error message(s)
-  getErrorsMessages: function (fieldInstance) {
+  getErrorsMessages: function () {
     // No error message, field is valid
-    if (true === fieldInstance.validationResult)
+    if (true === this.validationResult)
       return [];
 
     var messages = [];
 
-    for (var i = 0; i < fieldInstance.validationResult.length; i++)
-      messages.push(fieldInstance.validationResult[i].errorMessage ||
-       this._getErrorMessage(fieldInstance, fieldInstance.validationResult[i].assert));
+    for (var i = 0; i < this.validationResult.length; i++)
+      messages.push(this.validationResult[i].errorMessage ||
+       this._getErrorMessage(this.validationResult[i].assert));
 
     return messages;
   },
 
-  manageStatusClass: function (fieldInstance) {
-    if (fieldInstance.hasConstraints() && fieldInstance.needsValidation() && true === fieldInstance.validationResult)
-      this._successClass(fieldInstance);
-    else if (fieldInstance.validationResult.length > 0)
-      this._errorClass(fieldInstance);
+  _manageStatusClass: function () {
+    if (this.hasConstraints() && this.needsValidation() && true === this.validationResult)
+      this._successClass();
+    else if (this.validationResult.length > 0)
+      this._errorClass();
     else
-      this._resetClass(fieldInstance);
+      this._resetClass();
   },
 
-  manageErrorsMessages: function (fieldInstance, diff) {
-    if ('undefined' !== typeof fieldInstance.options.errorsMessagesDisabled)
+  _manageErrorsMessages: function (diff) {
+    if ('undefined' !== typeof this.options.errorsMessagesDisabled)
       return;
 
     // Case where we have errorMessage option that configure an unique field error message, regardless failing validators
-    if ('undefined' !== typeof fieldInstance.options.errorMessage) {
+    if ('undefined' !== typeof this.options.errorMessage) {
       if ((diff.added.length || diff.kept.length)) {
-        this._insertErrorWrapper(fieldInstance);
+        this._insertErrorWrapper();
 
-        if (0 === fieldInstance._ui.$errorsWrapper.find('.parsley-custom-error-message').length)
-          fieldInstance._ui.$errorsWrapper
+        if (0 === this._ui.$errorsWrapper.find('.parsley-custom-error-message').length)
+          this._ui.$errorsWrapper
             .append(
-              $(fieldInstance.options.errorTemplate)
+              $(this.options.errorTemplate)
               .addClass('parsley-custom-error-message')
             );
 
-        return fieldInstance._ui.$errorsWrapper
+        return this._ui.$errorsWrapper
           .addClass('filled')
           .find('.parsley-custom-error-message')
-          .html(fieldInstance.options.errorMessage);
+          .html(this.options.errorMessage);
       }
 
-      return fieldInstance._ui.$errorsWrapper
+      return this._ui.$errorsWrapper
         .removeClass('filled')
         .find('.parsley-custom-error-message')
         .remove();
@@ -98,299 +155,204 @@ ParsleyUI.prototype = {
 
     // Show, hide, update failing constraints messages
     for (var i = 0; i < diff.removed.length; i++)
-      this.removeError(fieldInstance, diff.removed[i].assert.name, true);
+      this.removeError(diff.removed[i].assert.name, {updateClass: false});
 
     for (i = 0; i < diff.added.length; i++)
-      this.addError(fieldInstance, diff.added[i].assert.name, diff.added[i].errorMessage, diff.added[i].assert, true);
+      this.addError(diff.added[i].assert.name, {message: diff.added[i].errorMessage, assert: diff.added[i].assert, updateClass: false});
 
     for (i = 0; i < diff.kept.length; i++)
-      this.updateError(fieldInstance, diff.kept[i].assert.name, diff.kept[i].errorMessage, diff.kept[i].assert, true);
+      this.updateError(diff.kept[i].assert.name, {message: diff.kept[i].errorMessage, assert: diff.kept[i].assert, updateClass: false});
   },
 
   // TODO: strange API here, intuitive for manual usage with addError(pslyInstance, 'foo', 'bar')
   // but a little bit complex for above internal usage, with forced undefined parameter...
-  addError: function (fieldInstance, name, message, assert, doNotUpdateClass) {
-    this._insertErrorWrapper(fieldInstance);
-    fieldInstance._ui.$errorsWrapper
+  addError: function (name, {message, assert, updateClass = true} = {}) {
+    this._insertErrorWrapper();
+    this._ui.$errorsWrapper
       .addClass('filled')
       .append(
-        $(fieldInstance.options.errorTemplate)
+        $(this.options.errorTemplate)
         .addClass('parsley-' + name)
-        .html(message || this._getErrorMessage(fieldInstance, assert))
+        .html(message || this._getErrorMessage(assert))
       );
 
-    if (true !== doNotUpdateClass)
-      this._errorClass(fieldInstance);
+    if (updateClass)
+      this._errorClass();
   },
 
   // Same as above
-  updateError: function (fieldInstance, name, message, assert, doNotUpdateClass) {
-    fieldInstance._ui.$errorsWrapper
+  updateError: function (name, {message, assert, updateClass = true} = {}) {
+    this._ui.$errorsWrapper
       .addClass('filled')
       .find('.parsley-' + name)
-      .html(message || this._getErrorMessage(fieldInstance, assert));
+      .html(message || this._getErrorMessage(assert));
 
-    if (true !== doNotUpdateClass)
-      this._errorClass(fieldInstance);
+    if (updateClass)
+      this._errorClass();
   },
 
   // Same as above twice
-  removeError: function (fieldInstance, name, doNotUpdateClass) {
-    fieldInstance._ui.$errorsWrapper
+  removeError: function (name, {updateClass = true} = {}) {
+    this._ui.$errorsWrapper
       .removeClass('filled')
       .find('.parsley-' + name)
       .remove();
 
-    // edge case possible here: remove a standard Parsley error that is still failing in fieldInstance.validationResult
+    // edge case possible here: remove a standard Parsley error that is still failing in this.validationResult
     // but highly improbable cuz' manually removing a well Parsley handled error makes no sense.
-    if (true !== doNotUpdateClass)
-      this.manageStatusClass(fieldInstance);
+    if (updateClass)
+      this._manageStatusClass();
   },
 
-  focus: function (formInstance) {
-    formInstance._focusedField = null;
-
-    if (true === formInstance.validationResult || 'none' === formInstance.options.focus)
-      return null;
-
-    for (var i = 0; i < formInstance.fields.length; i++) {
-      var field = formInstance.fields[i];
-      if (true !== field.validationResult && field.validationResult.length > 0 && 'undefined' === typeof field.options.noFocus) {
-        formInstance._focusedField = field.$element;
-        if ('first' === formInstance.options.focus)
-          break;
-      }
-    }
-
-    if (null === formInstance._focusedField)
-      return null;
-
-    return formInstance._focusedField.focus();
-  },
-
-  _getErrorMessage: function (fieldInstance, constraint) {
+  _getErrorMessage: function (constraint) {
     var customConstraintErrorMessage = constraint.name + 'Message';
 
-    if ('undefined' !== typeof fieldInstance.options[customConstraintErrorMessage])
-      return window.Parsley.formatMessage(fieldInstance.options[customConstraintErrorMessage], constraint.requirements);
+    if ('undefined' !== typeof this.options[customConstraintErrorMessage])
+      return window.Parsley.formatMessage(this.options[customConstraintErrorMessage], constraint.requirements);
 
     return window.Parsley.getErrorMessage(constraint);
   },
 
-  _diff: function (newResult, oldResult, deep) {
-    var added = [];
-    var kept = [];
-
-    for (var i = 0; i < newResult.length; i++) {
-      var found = false;
-
-      for (var j = 0; j < oldResult.length; j++)
-        if (newResult[i].assert.name === oldResult[j].assert.name) {
-          found = true;
-          break;
-        }
-
-      if (found)
-        kept.push(newResult[i]);
-      else
-        added.push(newResult[i]);
-    }
-
-    return {
-      kept: kept,
-      added: added,
-      removed: !deep ? this._diff(oldResult, newResult, true).added : []
-    };
-  },
-
-  setupForm: function (formInstance) {
-    formInstance.$element.on('submit.Parsley', evt => { formInstance.onSubmitValidate(evt); });
-    formInstance.$element.on('click.Parsley', 'input[type="submit"], button[type="submit"]', evt => { formInstance.onSubmitButton(evt); });
-
-    // UI could be disabled
-    if (false === formInstance.options.uiEnabled)
+  _buildUI: function () {
+    // UI could be already built or disabled
+    if (this._ui || false === this.options.uiEnabled)
       return;
 
-    formInstance.$element.attr('novalidate', '');
-  },
-
-  setupField: function (fieldInstance) {
-    var _ui = {active: false};
-
-    // UI could be disabled
-    if (false === fieldInstance.options.uiEnabled)
-      return;
-
-    _ui.active = true;
+    var _ui = {};
 
     // Give field its Parsley id in DOM
-    fieldInstance.$element.attr(fieldInstance.options.namespace + 'id', fieldInstance.__id__);
+    this.$element.attr(this.options.namespace + 'id', this.__id__);
 
-    /** Generate important UI elements and store them in fieldInstance **/
+    /** Generate important UI elements and store them in this **/
     // $errorClassHandler is the $element that woul have parsley-error and parsley-success classes
-    _ui.$errorClassHandler = this._manageClassHandler(fieldInstance);
+    _ui.$errorClassHandler = this._manageClassHandler();
 
     // $errorsWrapper is a div that would contain the various field errors, it will be appended into $errorsContainer
-    _ui.errorsWrapperId = 'parsley-id-' + (fieldInstance.options.multiple ? 'multiple-' + fieldInstance.options.multiple : fieldInstance.__id__);
-    _ui.$errorsWrapper = $(fieldInstance.options.errorsWrapper).attr('id', _ui.errorsWrapperId);
+    _ui.errorsWrapperId = 'parsley-id-' + (this.options.multiple ? 'multiple-' + this.options.multiple : this.__id__);
+    _ui.$errorsWrapper = $(this.options.errorsWrapper).attr('id', _ui.errorsWrapperId);
 
     // ValidationResult UI storage to detect what have changed bwt two validations, and update DOM accordingly
     _ui.lastValidationResult = [];
     _ui.validationInformationVisible = false;
 
-    // Store it in fieldInstance for later
-    fieldInstance._ui = _ui;
-
-    // Bind triggers first time
-    this.actualizeTriggers(fieldInstance);
+    // Store it in this for later
+    this._ui = _ui;
   },
 
   // Determine which element will have `parsley-error` and `parsley-success` classes
-  _manageClassHandler: function (fieldInstance) {
+  _manageClassHandler: function () {
     // An element selector could be passed through DOM with `data-parsley-class-handler=#foo`
-    if ('string' === typeof fieldInstance.options.classHandler && $(fieldInstance.options.classHandler).length)
-      return $(fieldInstance.options.classHandler);
+    if ('string' === typeof this.options.classHandler && $(this.options.classHandler).length)
+      return $(this.options.classHandler);
 
     // Class handled could also be determined by function given in Parsley options
-    var $handler = fieldInstance.options.classHandler(fieldInstance);
+    var $handler = this.options.classHandler.call(this, this);
 
     // If this function returned a valid existing DOM element, go for it
     if ('undefined' !== typeof $handler && $handler.length)
       return $handler;
 
     // Otherwise, if simple element (input, texatrea, select...) it will perfectly host the classes
-    if (!fieldInstance.options.multiple || fieldInstance.$element.is('select'))
-      return fieldInstance.$element;
+    if (!this.options.multiple || this.$element.is('select'))
+      return this.$element;
 
     // But if multiple element (radio, checkbox), that would be their parent
-    return fieldInstance.$element.parent();
+    return this.$element.parent();
   },
 
-  _insertErrorWrapper: function (fieldInstance) {
+  _insertErrorWrapper: function () {
     var $errorsContainer;
 
     // Nothing to do if already inserted
-    if (0 !== fieldInstance._ui.$errorsWrapper.parent().length)
-      return fieldInstance._ui.$errorsWrapper.parent();
+    if (0 !== this._ui.$errorsWrapper.parent().length)
+      return this._ui.$errorsWrapper.parent();
 
-    if ('string' === typeof fieldInstance.options.errorsContainer) {
-      if ($(fieldInstance.options.errorsContainer).length)
-        return $(fieldInstance.options.errorsContainer).append(fieldInstance._ui.$errorsWrapper);
+    if ('string' === typeof this.options.errorsContainer) {
+      if ($(this.options.errorsContainer).length)
+        return $(this.options.errorsContainer).append(this._ui.$errorsWrapper);
       else
-        ParsleyUtils.warn('The errors container `' + fieldInstance.options.errorsContainer + '` does not exist in DOM');
-    } else if ('function' === typeof fieldInstance.options.errorsContainer)
-      $errorsContainer = fieldInstance.options.errorsContainer(fieldInstance);
+        ParsleyUtils.warn('The errors container `' + this.options.errorsContainer + '` does not exist in DOM');
+    } else if ('function' === typeof this.options.errorsContainer)
+      $errorsContainer = this.options.errorsContainer.call(this, this);
 
     if ('undefined' !== typeof $errorsContainer && $errorsContainer.length)
-      return $errorsContainer.append(fieldInstance._ui.$errorsWrapper);
+      return $errorsContainer.append(this._ui.$errorsWrapper);
 
-    var $from = fieldInstance.$element;
-    if (fieldInstance.options.multiple)
+    var $from = this.$element;
+    if (this.options.multiple)
       $from = $from.parent();
-    return $from.after(fieldInstance._ui.$errorsWrapper);
+    return $from.after(this._ui.$errorsWrapper);
   },
 
-  actualizeTriggers: function (fieldInstance) {
-    var $toBind = fieldInstance._findRelated();
+  _actualizeTriggers: function () {
+    var $toBind = this._findRelated();
 
-    // Remove Parsley events already binded on this field
+    // Remove Parsley events already bound on this field
     $toBind.off('.Parsley');
-
-    // If no trigger is set, all good
-    if (false === fieldInstance.options.trigger)
-      return;
-
-    var triggers = fieldInstance.options.trigger.replace(/^\s+/g , '').replace(/\s+$/g , '');
-
-    if ('' === triggers)
-      return;
-
-    $toBind.on(
-      triggers.split(' ').join('.Parsley ') + '.Parsley',
-      event => { this.eventValidate(fieldInstance, event); }
-    );
+    if (this._failedOnce)
+      $toBind.on(ParsleyUtils.namespaceEvents(this.options.triggerAfterFailure, 'Parsley'), () => {
+        this.validate();
+      });
+    else {
+      $toBind.on(ParsleyUtils.namespaceEvents(this.options.trigger, 'Parsley'), event => {
+        this._eventValidate(event);
+      });
+    }
   },
 
-  eventValidate: function (field, event) {
-    // For keyup, keypress, keydown... events that could be a little bit obstrusive
+  _eventValidate: function (event) {
+    // For keyup, keypress, keydown, input... events that could be a little bit obstrusive
     // do not validate if val length < min threshold on first validation. Once field have been validated once and info
     // about success or failure have been displayed, always validate with this trigger to reflect every yalidation change.
-    if (/key/.test(event.type))
-      if (!field._ui.validationInformationVisible && field.getValue().length <= field.options.validationThreshold)
+    if (/key|input/.test(event.type))
+      if (!(this._ui && field._ui.validationInformationVisible) && this.getValue().length <= this.options.validationThreshold)
         return;
 
-    field.validate();
+    this.validate();
   },
 
-  manageFailingFieldTrigger: function (fieldInstance) {
-    fieldInstance._ui.failedOnce = true;
-
-    // Radio and checkboxes fields must bind every field multiple
-    if (fieldInstance.options.multiple)
-      fieldInstance._findRelated().each(function () {
-        if (!/change/i.test($(this).parsley().options.trigger || ''))
-          $(this).on('change.ParsleyFailedOnce', () => { fieldInstance.validate(); });
-      });
-
-    // Select case
-    if (fieldInstance.$element.is('select'))
-      if (!/change/i.test(fieldInstance.options.trigger || ''))
-        return fieldInstance.$element.on('change.ParsleyFailedOnce', () => { fieldInstance.validate(); });
-
-    // All other inputs fields
-    if (!/keyup/i.test(fieldInstance.options.trigger || ''))
-      return fieldInstance.$element.on('keyup.ParsleyFailedOnce', () => { fieldInstance.validate(); });
-  },
-
-  reset: function (parsleyInstance) {
+  _resetUI: function () {
     // Reset all event listeners
-    this.actualizeTriggers(parsleyInstance);
-    parsleyInstance.$element.off('.ParsleyFailedOnce');
+    this._failedOnce = false;
+    this._actualizeTriggers();
 
     // Nothing to do if UI never initialized for this field
-    if ('undefined' === typeof parsleyInstance._ui)
-      return;
-
-    if ('ParsleyForm' === parsleyInstance.__class__)
+    if ('undefined' === typeof this._ui)
       return;
 
     // Reset all errors' li
-    parsleyInstance._ui.$errorsWrapper
+    this._ui.$errorsWrapper
       .removeClass('filled')
       .children()
       .remove();
 
     // Reset validation class
-    this._resetClass(parsleyInstance);
+    this._resetClass();
 
     // Reset validation flags and last validation result
-    parsleyInstance._ui.lastValidationResult = [];
-    parsleyInstance._ui.validationInformationVisible = false;
-    parsleyInstance._ui.failedOnce = false;
+    this._ui.lastValidationResult = [];
+    this._ui.validationInformationVisible = false;
   },
 
-  destroy: function (parsleyInstance) {
-    this.reset(parsleyInstance);
+  _destroyUI: function () {
+    this._resetUI();
 
-    if ('ParsleyForm' === parsleyInstance.__class__)
-      return;
+    if ('undefined' !== typeof this._ui)
+      this._ui.$errorsWrapper.remove();
 
-    if ('undefined' !== typeof parsleyInstance._ui)
-      parsleyInstance._ui.$errorsWrapper.remove();
-
-    delete parsleyInstance._ui;
+    delete this._ui;
   },
 
-  _successClass: function (fieldInstance) {
-    fieldInstance._ui.validationInformationVisible = true;
-    fieldInstance._ui.$errorClassHandler.removeClass(fieldInstance.options.errorClass).addClass(fieldInstance.options.successClass);
+  _successClass: function () {
+    this._ui.validationInformationVisible = true;
+    this._ui.$errorClassHandler.removeClass(this.options.errorClass).addClass(this.options.successClass);
   },
-  _errorClass: function (fieldInstance) {
-    fieldInstance._ui.validationInformationVisible = true;
-    fieldInstance._ui.$errorClassHandler.removeClass(fieldInstance.options.successClass).addClass(fieldInstance.options.errorClass);
+  _errorClass: function () {
+    this._ui.validationInformationVisible = true;
+    this._ui.$errorClassHandler.removeClass(this.options.successClass).addClass(this.options.errorClass);
   },
-  _resetClass: function (fieldInstance) {
-    fieldInstance._ui.$errorClassHandler.removeClass(fieldInstance.options.successClass).removeClass(fieldInstance.options.errorClass);
+  _resetClass: function () {
+    this._ui.$errorClassHandler.removeClass(this.options.successClass).removeClass(this.options.errorClass);
   }
 };